rpi_data/modules/requirements.txt
pandas
bs4
requests
lxml
regex
pyyaml
